requirements.txt
flask
bs4
langchain
langchain-community
langchainhub
langchain-google-genai
langchain-pinecone
pandas
google-generativeai
pinecone-client
python-dotenv
textwrap3
ipywidgets
-e .